+++ app/code/community/Nosto/Tagging/Model/Meta/Product.php
@@ -225,6 +225,20 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Tagging_Model_Base implemen
     }
 
     /**
+     * Reloads the product info from a Magento product model.
+     *
+     * @param Mage_Catalog_Model_Product $product the product model to reload
+     * @param Mage_Core_Model_Store|null $store the store to get the product data for.
+     */
+    public function reloadData(Mage_Catalog_Model_Product $product, Mage_Core_Model_Store $store)
+    {
+        $reloadedProduct = Mage::getModel('catalog/product')
+            ->setStoreId($store->getId())
+            ->load($product->getId());
+        $this->loadData($reloadedProduct, $store);
+    }
+
+    /**
      * Loads the product info from a Magento product model.
      *
      * @param Mage_Catalog_Model_Product $product the product model.